Resolves issue #449 and improves the organization of city and civ modifiers. Also extends CivModifier so it's capable of doing more things.

+++ Assets/Simulation/Improvements/IImprovement.cs
@@ -19,7 +19,7 @@ namespace Assets.Simulation.Improvements {
 
         bool IsReadyToConstruct { get; }
 
-        int WorkInvested { get; set; }
+        float WorkInvested { get; set; }
 
         Transform transform { get; }
